+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 8045dbd - Update DAW/BAE/Unidad-2/Tarea-6
+| Author: fjrodafo - Sun, 16 Feb 2025 20:13:41 +0000
+| 
 * Commit: 6ee19f1 - Update workflows runner to ubuntu-latest
 | Author: fjrodafo - Sun, 16 Feb 2025 18:45:43 +0000
 |